Full restructuring of profile
