+++ lib/features/profile/presentation/info_body/info_body_adapter.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+abstract interface class InfoBody extends Widget {
+  const InfoBody({super.key});
+}
