+++ lib/features/profile/presentation/profile_body.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:gymbro/core/_dev/debug_tools.dart';
-import 'gallery_block.dart';
-import 'entries_block.dart';
-import 'profile_info.dart';
+import 'package:gymbro/features/profile/presentation/profile_header.dart';
+import 'info_body/gallery_block.dart';
+import 'info_body/entries_block.dart';
+import 'info_body/info_wrapper.dart';
 import 'package:gymbro/core/widgets/double_text.dart';
 
 class ProfileBody extends StatelessWidget {
@@ -36,7 +37,7 @@ class ProfileBody extends StatelessWidget {
             ],
           ),
           // Basically, just need a "ручка"
-          InfoSection(
+          InfoWrapper(
             header: 'General Info',
             optionalButton: null,
             infoBody: Entries(
@@ -47,104 +48,18 @@ class ProfileBody extends StatelessWidget {
               ],
             )
           ),
-          InfoSection(
+          InfoWrapper(
             header: 'Progress in Pictures',
             infoBody: Gallery(
               photosUrls: [
-                'assets/dog.png',
-                'assets/cat.png',
-                'assets/myles.jpg',
+                'assets/images/dog.jpeg',
+                'assets/images/cat.jpeg',
+                'assets/images/myles.jpeg',
               ],
             ),
           ),
         ]
       ),
-    );
-  }
-}
-
-class ProfileHeader extends StatelessWidget {
-  const ProfileHeader({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final topRelPadding = ProfileBody.topRelPadding;
-    final radius = ProfileBody.radius;
-    final leftPadding = ProfileBody.leftPadding;
-    
-    final pbgHeight = ProfileBody.pbgHeight;
-
-    final double nameLeftPadding = ((0.8 - topRelPadding) * radius / 2).abs();
-
-    final ThemeData contextTheme = Theme.of(context); 
-
-    return Positioned(
-      left: leftPadding,
-      top: pbgHeight - radius * topRelPadding,
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.end,
-        children: [
-          ProfileAvatar(),
-          DebugGizmo(
-            child: Padding(
-              padding: EdgeInsets.only(left: nameLeftPadding),
-              child: Column(
-                children: [
-                  DoubleTextDisplay(
-                    topText: 'Name LastName', 
-                    topStyle: contextTheme.textTheme.titleLarge,
-                    bottomText: 'online',
-                    bottomColor: Colors.grey,
-                  ),
-                ],
-              ),
-            ),
-          ),
-          
-        ],
-      ),
-    );
-  }
-}
-
-class ProfileAvatar extends StatelessWidget {
-  const ProfileAvatar({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final radius = ProfileBody.radius;
-    return ClipOval(
-        child: Image(
-          width: radius,
-          height: radius,
-          image: NetworkImage('https://flutter.github.io/assets-for-api-docs/assets/widgets/owl.jpg'),
-        ),
-    );
-  }
-}
-
-class ProfileBackGround extends StatelessWidget {
-  const ProfileBackGround({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final double pbgHeight = ProfileBody.pbgHeight;
-    return Row(
-      children: [
-        Expanded(
-          child: ColoredBox(
-            color: Theme.of(context).colorScheme.primary,
-            child: SizedBox(height: pbgHeight,),
-          ),
-        ),
-      ],
     );
   }
 }
