+++ lib/features/profile/presentation/profile_page.dart
@@ -1,6 +1,8 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'settings_button.dart';
+import 'info_body/header_buttons/settings_button.dart';
 import 'package:gymbro/core/_dev/debug_tools.dart';
 import 'profile_body.dart';
 
@@ -23,7 +25,7 @@ class ProfilePage extends ConsumerWidget {
         actions: [
           SettingsMenu(),
         ],
-        title: Text('Profile'),
+        title: Text(""),
       ),
       body: ProfileBody(),
     );
